+++ MoltenVK/MoltenVK/Commands/MVKCommandBuffer.mm
@@ -292,7 +292,12 @@ void MVKCommandEncoder::beginMetalRenderPass(bool loadOverride, bool storeOverri
     getSubpass()->populateMTLRenderPassDescriptor(mtlRPDesc, _framebuffer, _clearValues, _isRenderingEntireAttachment, loadOverride, storeOverride);
     mtlRPDesc.visibilityResultBuffer = _occlusionQueryState.getVisibilityResultMTLBuffer();
 
-	if (_device->_pMetalFeatures->layeredRendering) {
+	// Only set the layered rendering properties if layered rendering is supported and the framebuffer really has multiple layers
+	if ((_framebuffer->getLayerCount() > 1) &&
+		_device->_pMetalFeatures->layeredRendering &&
+		(_device->_pMetalFeatures->multisampleLayeredRendering ||
+		 (getSubpass()->getSampleCount() == VK_SAMPLE_COUNT_1_BIT))) {
+
 		VkExtent2D fbExtent = _framebuffer->getExtent2D();
 		mtlRPDesc.renderTargetWidthMVK = min(_renderArea.offset.x + _renderArea.extent.width, fbExtent.width);
 		mtlRPDesc.renderTargetHeightMVK = min(_renderArea.offset.y + _renderArea.extent.height, fbExtent.height);